requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
aiofiles>=23.2.1
requests>=2.31.0
python-dotenv>=1.0.0
deepagents==0.3.0
langchain-openai>=0.1.0
gunicorn>=21.2.0
